feat(ElTableRem): 新增 el-table-col-rem组件
修正传入width之后，parseWidth不能转换小数的问题

+++ packages/table-rem/src/util.js
@@ -152,7 +152,7 @@ export function mergeOptions(defaults, config) {
 
 export function parseWidth(width) {
   if (width !== undefined) {
-    width = parseInt(width, 10);
+    width = parseFloat(width, 10);
     if (isNaN(width)) {
       width = null;
     }
